@@ -1,72 +1,75 @@
 {
-  "name": "psyops-svelte",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "dev": "  vite dev",
-    "build": "vite build",
-    "preview": "vite preview",
-    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-    "lint": "prettier --check .",
-    "format": "prettier --write .",
-    "i18n": "typesafe-i18n"
-  },
-  "devDependencies": {
-    "@iconify-json/bi": "^1.2.1",
-    "@iconify-json/flat-color-icons": "^1.2.0",
-    "@iconify-json/fluent-color": "^1.2.1",
-    "@iconify-json/fluent-emoji": "^1.2.1",
-    "@iconify-json/logos": "^1.2.3",
-    "@iconify-json/mdi": "^1.2.1",
-    "@inlang/paraglide-js": "1.11.3",
-    "@poppanator/sveltekit-svg": "^5.0.0",
-    "@sveltejs/adapter-auto": "3.3.1",
-    "@sveltejs/kit": "2.8.0",
-    "@sveltejs/vite-plugin-svelte": "4.0.0",
-    "@types/better-sqlite3": "7.6.11",
-    "@types/bun": "1.1.13",
-    "@vite-pwa/sveltekit": "^0.6.6",
-    "autoprefixer": "10.4.20",
-    "daisyui": "4.12.14",
-    "editorjs-inline-spoiler-tool": "1.0.1",
-    "postcss": "8.4.49",
-    "prettier": "3.3.3",
-    "prettier-plugin-svelte": "3.2.8",
-    "svelte": "5.1.16",
-    "svelte-check": "4.0.7",
-    "tailwindcss": "3.4.14",
-    "typesafe-i18n": "5.26.2",
-    "typescript": "5.6.3",
-    "unplugin-icons": "^0.20.1",
-    "vite": "5.4.11"
-  },
-  "type": "module",
-  "dependencies": {
-    "@editorjs/editorjs": "2.30.7",
-    "@editorjs/marker": "1.4.0",
-    "@iconify/svelte": "4.0.2",
-    "@inlang/paraglide-sveltekit": "0.11.5",
-    "@oslojs/binary": "1.0.0",
-    "@oslojs/crypto": "1.0.1",
-    "@oslojs/encoding": "1.1.0",
-    "@pilcrowjs/db-query": "0.0.2",
-    "@pilcrowjs/object-parser": "0.0.4",
-    "@svitejs/svelte-preprocess-svg": "2.0.0",
-    "@tiptap/core": "^2.9.1",
-    "@tiptap/pm": "^2.9.1",
-    "@tiptap/starter-kit": "^2.9.1",
-    "arctic": "2.2.2",
-    "better-sqlite3": "11.5.0",
-    "editorjs-table": "1.4.10",
-    "editorjs-telegram-post": "1.0.6",
-    "editorjs-undo": "2.0.28",
-    "editorjs-youtube-embed": "1.0.3",
-    "novel-svelte": "^0.1.7",
-    "panzoom": "^9.4.3",
-    "postgres": "^3.4.5",
-    "simple-image-editorjs": "1.4.0",
-    "svelte-editorjs": "0.1.0",
-    "svelte-tiptap": "^2.0.3"
-  }
+	"name": "psyops-svelte",
+	"version": "0.0.1",
+	"private": true,
+	"scripts": {
+		"dev": "vite dev",
+		"build": "vite build",
+		"preview": "vite preview",
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+		"lint": "prettier --check .",
+		"format": "prettier --write .",
+		"i18n": "typesafe-i18n"
+	},
+	"devDependencies": {
+		"@iconify-json/bi": "1.2.1",
+		"@iconify-json/flat-color-icons": "1.2.0",
+		"@iconify-json/fluent-color": "1.2.1",
+		"@iconify-json/fluent-emoji": "1.2.1",
+		"@iconify-json/logos": "1.2.3",
+		"@iconify-json/mdi": "1.2.1",
+		"@inlang/paraglide-js": "1.11.3",
+		"@poppanator/sveltekit-svg": "5.0.0",
+		"@sveltejs/adapter-auto": "3.3.1",
+		"@sveltejs/kit": "2.9.0",
+		"@sveltejs/vite-plugin-svelte": "5.0.1",
+		"@types/better-sqlite3": "7.6.12",
+		"@types/bun": "1.1.14",
+		"@vite-pwa/sveltekit": "0.6.6",
+		"daisyui": "4.12.14",
+		"editorjs-inline-spoiler-tool": "1.0.1",
+		"prettier": "3.4.1",
+		"prettier-plugin-svelte": "3.3.2",
+		"svelte": "5.2.11",
+		"svelte-check": "4.1.0",
+		"tailwindcss": "3.4.15",
+		"typesafe-i18n": "5.26.2",
+		"typescript": "5.7.2",
+		"unplugin-icons": "0.20.2",
+		"vite": "6.0.1"
+	},
+	"type": "module",
+	"dependencies": {
+		"@auth/pg-adapter": "1.7.4",
+		"@auth/sveltekit": "1.7.4",
+		"@editorjs/editorjs": "2.30.7",
+		"@editorjs/marker": "1.4.0",
+		"@iconify/svelte": "4.0.2",
+		"@inlang/paraglide-sveltekit": "0.12.1",
+		"@oslojs/binary": "1.0.0",
+		"@oslojs/crypto": "1.0.1",
+		"@oslojs/encoding": "1.1.0",
+		"@pilcrowjs/db-query": "0.0.2",
+		"@pilcrowjs/object-parser": "0.0.4",
+		"@svitejs/svelte-preprocess-svg": "2.0.0",
+		"@tailwindcss/vite": "^4.0.0-beta.3",
+		"@tiptap/core": "2.10.3",
+		"@tiptap/pm": "2.10.3",
+		"@tiptap/starter-kit": "2.10.3",
+		"arctic": "2.3.0",
+		"better-sqlite3": "11.6.0",
+		"editorjs-table": "1.4.10",
+		"editorjs-telegram-post": "1.0.6",
+		"editorjs-undo": "2.0.28",
+		"editorjs-youtube-embed": "1.0.3",
+		"novel-svelte": "0.1.7",
+		"panzoom": "9.4.3",
+		"pg": "8.13.1",
+		"postgres": "3.4.5",
+		"simple-image-editorjs": "1.4.0",
+		"svelte-editorjs": "0.1.0",
+		"svelte-tiptap": "2.0.3",
+		"uuid": "11.0.3"
+	}
 }
