@@ -1,8 +1,8 @@
 <script lang="ts">
+	import "../app.css";
 	import { ParaglideJS } from "@inlang/paraglide-sveltekit";
 	import { i18n } from "$lib/i18n";
 
-	import "../app.css";
 	import { goto } from "$app/navigation";
 	import { page, navigating } from "$app/stores";
 
