@@ -5,13 +5,15 @@
 	import FluentColorGlobeShield24 from "~icons/fluent-color/globe-shield-24";
 </script>
 
-<main class="flex flex-col items-center justify-center h-full p-4 text-center">
+<main
+	class="flex flex-col items-center justify-center h-full p-4 text-center bg-gradient-to-r from-purple-800 to-zinc-800"
+>
 	<div
 		class="flex flex-col items-center justify-center gap-1 p-4 text-center border rounded-lg border-fuchsia-900 place-self-center"
 	>
 		<FluentColorGlobeShield24 class="w-12 h-12" />
 
-		<a href="/login/google" class="mt-4 btn btn-primary btn-wide">
+		<a href="/login/google" class="mt-4 btn btn-primary btn-wide glass">
 			<LogosGoogleIcon class="w-5 h-5 " />
 			{m.signUp({ provider: "Google" })}
 		</a>
